erc20 indexing working
remove unused code
