+++ .gitignore
@@ -17,4 +17,4 @@ public/
 
 subgraph.yaml
 
-src/mappings/projectTokens
+src/erc20/**
