+++ subgraph.template.yaml
@@ -4,25 +4,6 @@ features:
   - fullTextSearch
 schema:
   file: ./schema.graphql
-templates:
-  - name: TreasuryToken
-    kind: ethereum/contract
-    network: {{network}}
-    source:
-      abi: ERC20
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.5
-      language: wasm/assemblyscript
-      file: ./src/mappings/erc20.ts
-      entities:
-        - Exchange
-      abis:
-        - name: ERC20
-          file: ./abis/ERC20.json
-      eventHandlers:
-        - event: Transfer(indexed address,indexed address,uint256)
-          handler: handleTransfer
 dataSources:
   - kind: ethereum/contract
     name: Projects
@@ -121,6 +102,7 @@ dataSources:
         - event: Migrate(indexed uint256,indexed address,uint256,address)
           handler: handleMigrate
       file: ./src/mappings/terminalV1.ts
+  {{ #address_terminalV1_1 }}
   - kind: ethereum/contract
     name: TerminalV1_1
     network: {{network}}
@@ -164,9 +146,8 @@ dataSources:
         - event: DistributeToTicketMod(indexed uint256,indexed uint256,(bool,uint16,uint48,address),uint256,address)
           handler: handleDistributeToTicketMod
       file: ./src/mappings/terminalV1_1.ts
-  # Add all token contract references with address
-  {{ #projectTokens }}
-  {{ #address  }}
+  {{ /address_terminalV1_1 }}
+  {{ #erc20s }}
   - kind: ethereum/contract
     name: {{name}}
     network: {{network}}
@@ -186,6 +167,5 @@ dataSources:
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handle{{name}}Transfer
-      file: ./src/mappings/projectTokens/{{network}}.ts
-  {{ /address }}
-  {{ /projectTokens }}
+      file: ./src/erc20/mapping.ts
+  {{ /erc20s }}
